Move App to kapture-backend directory
